0.0.22.alpha1 prerelease
This is a short term release. I expect to remove this tag at some point after
I cut the next release.

The commits between 0.0.21 on main and this commit are likely to be rewritten,
as this is not a finalized changeset.

I may drop features between now and the next release, but this is quite unlikely.

I expect to leave this release up for some period after the next official
release.

Key features:
- Support for languages other than English, see:
  https://github.com/check-spelling/check-spelling/wiki/Feature%3A-Configurable-word-characters

- Exclude suggestions can be paths, filenames, file extensions:
  https://github.com/check-spelling/check-spelling/wiki/Feature:-Heuristic-exclude-suggestions#advanced-exclusion-suggestions

- Dictionary suggestions favor coverage
  https://github.com/check-spelling/check-spelling/wiki/Feature%3A-Improve-suggested-dictionaries-order#improve-suggested-dictionaries-order

- New Kubernetes dictionary
  https://github.com/check-spelling/check-spelling/wiki/Dictionary:-K8s

- New Lisp dictionary
  https://github.com/check-spelling/check-spelling/wiki/Dictionary:-Lisp

Breaking changes:
- The default dictionaries no longer include cpp.txt and win32.txt
  they included way too many typos to be used as a dictionary.
  https://github.com/check-spelling/check-spelling/wiki/Removed-Dictionaries#v0022